refactor: styling for stores

+++ main/models.py
@@ -10,7 +10,7 @@ class User(AbstractUser):
     ]
 
     role = models.CharField(max_length=20, choices=ROLE_CHOICES, default='normal_user')
-    money = models.IntegerField(default=0)  # Add the money field
+    money = models.DecimalField(default=0, max_digits=10, decimal_places=2)  # Add the money field
     created_at = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
